Update performance-linux-filesystem-cache.md
Please be careful with principal write influencing settings.
The ratio's are taken from available memory, not total memory, and thus get smaller as memory pressure increases. Lots of people get confused because they assume this is taken from total memory, and does not behave as they think for that reason.